added navbar and filmlist component, fixed form names
