+++ frontend/src/components/HamburgerMenu/HamburgerMenu.scss
@@ -1,0 +1,80 @@
+@import url("https://fonts.googleapis.com/css2?family=Roboto&display=swap");
+@import "../../styles/variables.scss";
+
+.hamburger-menu {
+    position: relative;
+    &__btn {
+      display: flex;
+      align-items: center;
+      position: absolute;
+      bottom: 0;
+      left: 0;
+      width: 26px;
+      height: 26px;
+      cursor: pointer;
+      z-index: 1;
+    }
+    &__box {
+      display: block;
+      position: fixed;
+      visibility: hidden;
+      top: 0;
+      left: -100%;
+      width: 75%;
+      height: 100%;
+      margin: 0;
+      padding: 80px 0;
+      list-style: none;
+      background-color: $activbtn;
+      box-shadow: 1px 0px 6px rgba(0, 0, 0, 0.2);
+      transition-duration: 0.25s;
+    }
+    &__item {
+      display: block;
+      padding: 12px 24px;
+      color: #fff;
+      font-size: 20px;
+      text-decoration: none;
+      transition-duration: 0.25s;
+    }
+  }
+  
+  #menu__toggle {
+    opacity: 0;
+  }
+  
+  #menu__toggle:checked ~ .hamburger-menu__btn > span {
+    transform: rotate(45deg);
+  }
+  #menu__toggle:checked ~ .hamburger-menu__btn > span::before {
+    top: 0;
+    transform: rotate(0);
+  }
+  #menu__toggle:checked ~ .hamburger-menu__btn > span::after {
+    top: 0;
+    transform: rotate(90deg);
+  }
+  #menu__toggle:checked ~ .hamburger-menu__box {
+    visibility: visible;
+    left: 0;
+  }
+
+  .hamburger-menu__btn > span,
+  .hamburger-menu__btn > span::before,
+  .hamburger-menu__btn > span::after {
+    display: block;
+    position: absolute;
+    width: 100%;
+    height: 3px;
+    background-color: #000;
+    transition-duration: 0.25s;
+  }
+  .hamburger-menu__btn > span::before {
+    content: "";
+    top: -8px;
+  }
+  .hamburger-menu__btn > span::after {
+    content: "";
+    top: 8px;
+  }
+
