+++ frontend/src/pages/Catalog.jsx
@@ -1,8 +1,13 @@
 import React from "react";
+import Navbar from "../components/Navbar/Navbar";
 
 const Catalog = (props) => {
-    return <div>Catalog</div>
-
-}
+  return (
+    <div>
+      <Navbar />
+      <div>Catalog</div>
+    </div>
+  );
+};
 
 export default Catalog;
