+++ frontend/src/pages/FilmLists/FilmLists.scss
@@ -1,0 +1,10 @@
+.filmLists {
+  &__title {
+    margin: 1rem;
+    text-transform: uppercase;
+    text-align: center;
+    font-size: 28px;
+    font-family: 'Roboto', sans-serif;
+    font-weight: 700;
+  }
+}
