+++ frontend/src/pages/RegistrationPage/RegistrationPage.jsx
@@ -1,0 +1,11 @@
+import React from "react"
+import RegistrationForm from "../../components/Forms/RegistrationForm";
+import "./RegistrationPage.scss";
+
+const RegistrationPage = (props) => {
+  return <RegistrationForm />
+}
+
+
+
+export default RegistrationPage;
