+++ frontend/src/pages/RegistrationPage/RegistrationPage.scss
