+++ frontend/src/styles/styles.scss
@@ -2,7 +2,7 @@ body {
   margin: 0;
   padding: 0;
   font-family: Arial, Helvetica, sans-serif;
-  transition: .2s all ease-in;
+  transition: 0.2s all ease-in;
   &.dark {
     background: #343a40;
     color: #fff;
