+++ frontend/src/styles/variables.scss
